chore: refactor with internal
